raisesgroup followups * renames src/_pytest/raises_group.py to src/_pytest/raises.py * moves pytest.raises from src/_pytest/python_api.py to src/_pytest/raises.py * adds several newsfragments that should've been bundlen with * add RaisesGroup & Matcher #13192 * add more detailed error message if you try to do RaisesGroup((ValueError, TypeError)) * mess around with ValueError vs TypeError on invalid expected exception
