+++ changelog/13192.feature.2.rst
@@ -1,0 +1,1 @@
+The context-manager form of :func:`pytest.raises` now has a ``check`` parameter that takes a callable that should return `True` when passed the raised exception if it should be accepted.
